+++ test/turbo_overlay/view_helper_test.rb
@@ -2,15 +2,14 @@ require "test_helper"
 require "turbo_overlay/helpers/view_helper"
 
 class ViewHelperTest < Minitest::Test
-  # Stand-in view context. Captures arguments passed to link_to /
-  # content_for so we can assert what the helpers forward.
   class FakeView
     include TurboOverlay::Helpers::ViewHelper
 
-    attr_reader :link_to_args, :content_for_calls, :modal_request_value
+    attr_reader :link_to_args, :content_for_calls, :modal_request_value, :drawer_request_value
 
-    def initialize(modal_request: false)
-      @modal_request_value = modal_request
+    def initialize(modal_request: false, drawer_request: false)
+      @modal_request_value  = modal_request
+      @drawer_request_value = drawer_request
       @link_to_args = nil
       @content_for_calls = []
     end
@@ -24,8 +23,20 @@ class ViewHelperTest < Minitest::Test
       @content_for_calls << [name, value, block]
     end
 
+    def turbo_frame_tag(id)
+      %(<turbo-frame id="#{id}"></turbo-frame>)
+    end
+
+    def safe_join(parts)
+      parts.join
+    end
+
     def modal_request?
       @modal_request_value
+    end
+
+    def drawer_request?
+      @drawer_request_value
     end
 
     def controller
@@ -52,8 +63,8 @@ class ViewHelperTest < Minitest::Test
     view.modal_link_to("Open", "/things/1")
 
     name, options, html_options = view.link_to_args
-    assert_equal "Open",       name
-    assert_equal "/things/1",  options
+    assert_equal "Open",      name
+    assert_equal "/things/1", options
     assert_equal({ turbo_frame: "turbo_modal" }, html_options[:data])
   end
 
@@ -92,6 +103,72 @@ class ViewHelperTest < Minitest::Test
     _, _, html_options = view.link_to_args
     refute html_options.key?("data-action")
     refute html_options.key?("data-turbo-modal-dismiss")
+  end
+
+  # ---- drawer_link_to ----
+
+  def test_drawer_link_to_adds_turbo_frame_data
+    view = FakeView.new
+    view.drawer_link_to("Filter", "/filters")
+
+    _, _, html_options = view.link_to_args
+    assert_equal({ turbo_frame: "turbo_drawer" }, html_options[:data])
+  end
+
+  def test_drawer_link_to_uses_configured_frame_id
+    TurboOverlay.configure { |c| c.drawer { |d| d.frame_id = "my_drawer" } }
+    view = FakeView.new
+    view.drawer_link_to("Filter", "/filters")
+
+    _, _, html_options = view.link_to_args
+    assert_equal({ turbo_frame: "my_drawer" }, html_options[:data])
+  end
+
+  # ---- drawer_dismiss_link_to ----
+
+  def test_drawer_dismiss_link_to_inside_drawer_adds_dismiss_action
+    view = FakeView.new(drawer_request: true)
+    view.drawer_dismiss_link_to("Close", "/back")
+
+    _, _, html_options = view.link_to_args
+    assert_includes html_options["data-action"], "click->turbo-drawer#close"
+    assert_equal "true", html_options["data-turbo-drawer-dismiss"]
+  end
+
+  def test_drawer_dismiss_link_to_outside_drawer_is_plain_link
+    view = FakeView.new(drawer_request: false)
+    view.drawer_dismiss_link_to("Close", "/back")
+
+    _, _, html_options = view.link_to_args
+    refute html_options.key?("data-action")
+    refute html_options.key?("data-turbo-drawer-dismiss")
+  end
+
+  # ---- overlay_frame_tags ----
+
+  def test_overlay_frame_tags_emits_both_by_default
+    view = FakeView.new
+    output = view.overlay_frame_tags
+    assert_includes output, %(<turbo-frame id="turbo_modal">)
+    assert_includes output, %(<turbo-frame id="turbo_drawer">)
+  end
+
+  def test_overlay_frame_tags_can_be_filtered
+    view = FakeView.new
+    output = view.overlay_frame_tags(:modal)
+    assert_includes output, %(<turbo-frame id="turbo_modal">)
+    refute_includes output, %(<turbo-frame id="turbo_drawer">)
+  end
+
+  def test_overlay_frame_tags_uses_configured_frame_ids
+    TurboOverlay.configure do |c|
+      c.modal  { |m| m.frame_id = "my_modal" }
+      c.drawer { |d| d.frame_id = "my_drawer" }
+    end
+    view = FakeView.new
+    output = view.overlay_frame_tags
+    assert_includes output, %(<turbo-frame id="my_modal">)
+    assert_includes output, %(<turbo-frame id="my_drawer">)
   end
 
   # ---- generic in-view content helpers ----